Displayed Person List
Replaced the template application with a single page of
Person List. The Person List if able to fetch the list from the
server and display it.

At the moment only this page can be displayed and the
data structure (model) is still the one experimentally exposed by the server.

The project got a new dependency - MaterialUI. The components are used
from this component framework. This will help to save some time in the
future.