Person API is returning API entity instead of database entity.
The PersonService is responsible to turn the database entity into a
API entity. The conversion itself s done in a separate object, in the
PersonConverter.

TODO: The PersonConverter is not fully done yet. It needs to convert
other values
TODO: The PersonService must propagate other values, like Family,
but it needs to be loaded first by the FamilyService (to be created)