Created endpoint to retrieve families
In order to provide the endpoint the Resource (Controller),
the Service and the Repository had to be created. There are also model
files for both the API and the DB model. The FamilyConverter is
to convert the DB Model into API Model.

As Gramps stores the data as pickles python object, those need to
be `unpickled` and then parsed in the proper order.