Dockerized the environment (for development)
Dockerized the environment for development.
Both frontend (web) and backend (api) are in docker container
frontend with a reverse proxy (nginx).

All request can go to http://localhost:8081, so we don't need
to worry bout CORS any more.

GraphiQL UI is also exposed on the http://localhost:8081/graphiql
url.

Note: the frontend is not working yet corretly. It needs to be
fixed.