Added Home page and React Router
Add React Router in order to be able to manage the navigation routes.
I also added a Home.tsx as the default landing page. It might
change in the future once authentication is added to the system.

The Persons.tsx page is not behind the usr /people and the main
nav bar has a button navigating there.