First version of Person Details display and form
An attempt to display person details with and editable form.
For the form I used `formik`.

`Person` page is hooked up to the `/people/{grampsId}` router path
and it tries to retrieve a person via the graphql api.
Once the person is there, the `PersonDetails` component is displayed
that shows the details of the person (currently only very basic data).

`NameField` is an editable namefield, that can be hooked up to `formik` forms.
`PersonDetails` is using a `formik` form to handle change to the data.

There is no solution to save the change (yet). It requires backend change
first and it needs to be supported by the graphql api.