PersonTree page to display around the given person
The PersonTree displays a small tree that shows the parents
and the children of the given person.

This is the first version of this component and it needs
further refinement before it becomes usable.

The new component and page is mounted on `/people/tree/{grampsId}`.
It limits the ability to retrieve a person with the id: `tree`,
but I think we can accept that tradeoff.

The tree can be later used for the photobook component.