Very first attempt to create a Photobook
The PhotoBook is in the `PhotoBook` page component,
and makes use the `PhotoBookPage` component, that
renders currently family photos on the left hand side
and the tree on the right hand side.

To render the photos onto the page, there is a piece of
small alogrithm that calculates the necessery row (or image)
height so the flex layout will fit onto the page.
This algorithm is not using a clear lingo and it needs to be
cleaned up. Also not sure if some values should be memoized there.

The print layout is also not implemented yet. Currently it only
works with the predefined A4 pagesize, but is should be
configurable.

A pages should be changed as well. If there are no images of
the family, the image page should be taken away.

It's a bit tricky because the photo and the tree page should
be always on the opposite sites and the structure of the tree
should not be violated (too much) either.

Other todos:
- The whole book should be a cover page.
- The book should use a different kind of font
- The number of generations should be configurable
- Add background

Possible extensions:
- Configure potential side-branches of the tree to be included into the photo book. (eg. cousin's family)
- Configurable background