PhotoBook cleanup
The different pages have been moved into their own component. The Tree is in the
`photobook/FamilyTreePage` whilst the Photos page is int the `photobook/PhotosPage`
component.

The responsibiliy to iterate over the images have been moved from `pages/PhotoBook`
page component into `PhotoBookPage` component.

TODO: the naming is still not fully correct. The goal of this commit was
to separate the code.